Changed VS Code extension
I was using ccls and the 'ccls-project.ccls' VS Code extension for a C
LSP client, however, that extension lacks of features, support/updates,
VS Code integration and is a bit tedious to setting up. But the C/C++
official VS Code extension by Microsoft is far better, the absolute
opposite. So, consequently, in this commit I:

  - Remove the .gitignore since is no longer required
  - Remove the .vscode/tasks.json since is no longer required
  - Added .vscode/settings.json as a workspace-specific settings for the
    new extension to work properly